add the auxn implementation
hi! I've been working on a reimplementation of uxn in rust with the added feature of it being able to run as a VST3 or CLAP audio plugin, i added it to the misc. section altough im not sure if that category is right. maybe should i move it to desktop? thanks